# openapi.json
{
  "components": {
    "schemas": {
...
      "FlockAcquisitionCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "FlockAcquisitionCreate",
        "type": "object"
      },
      "FlockActionRead": {
        "description": "The events a flock action emitted, plus the resulting on-hand headcount.\n\n``paired_event_id`` is the expense (acquisition), income (sale), or\nmortality (mortality) event \u2014 null only for an acquisition with no amount.",
        "properties": {
          "headcount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Headcount",
            "type": "string"
          },
          "inventory_event_id": {
            "title": "Inventory Event Id",
            "type": "integer"
          },
          "paired_event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paired Event Id"
          }
        },
        "required": [
          "inventory_event_id",
          "paired_event_id",
          "headcount"
        ],
        "title": "FlockActionRead",
        "type": "object"
      },
      "FlockMortalityCreate": {
        "additionalProperties": false,
        "properties": {
          "cause": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "FlockMortalityCreate",
        "type": "object"
      },
      "FlockSaleCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "buyer": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "quantity",
          "amount"
        ],
        "title": "FlockSaleCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaterialPurchaseCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "material_asset_id": {
            "title": "Material Asset Id",
            "type": "integer"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "supplier": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          }
        },
        "required": [
          "material_asset_id",
          "occurred_at",
          "quantity",
          "unit",
          "amount"
        ],
        "title": "MaterialPurchaseCreate",
        "type": "object"
      },
      "MaterialPurchaseRead": {
        "properties": {
          "amount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "expense_event_id": {
            "title": "Expense Event Id",
            "type": "integer"
          },
          "farm_id": {
            "title": "Farm Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "inventory_event_id": {
            "title": "Inventory Event Id",
            "type": "integer"
          },
          "material_asset_id": {
            "title": "Material Asset Id",
            "type": "integer"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "supplier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "farm_id",
          "material_asset_id",
          "inventory_event_id",
          "expense_event_id",
          "occurred_at",
          "quantity",
          "unit",
          "amount",
          "currency",
          "supplier",
          "notes",
          "meta",
          "idempotency_key",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "MaterialPurchaseRead",
        "type": "object"
      },
      "MaterialPurchaseUpdate": {
        "additionalProperties": false,
        "description": "PATCH body. ``material_asset_id`` is immutable \u2014 a different material is a\ndifferent purchase.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "supplier": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "MaterialPurchaseUpdate",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "has_next"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "Page_MaterialPurchaseRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MaterialPurchaseRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Page[MaterialPurchaseRead]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Ovejitas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/farms/{farm_id}/assets/{asset_id}/flock/acquisitions": {
      "post": {
        "description": "Increments the flock headcount by `quantity` head via an INVENTORY event; when `amount` is given, also books a paired EXPENSE in the farm's default currency. The asset must be `animal` + `aggregated`.",
        "operationId": "flock_create_acquisition",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlockAcquisitionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlockActionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Record a flock acquisition",
        "tags": [
          "flock"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/assets/{asset_id}/flock/mortalities": {
      "post": {
        "description": "Decrements the flock headcount by `quantity` head and emits a paired MORTALITY event. Rejected with 409 if it would drive the headcount below zero.",
        "operationId": "flock_create_mortality",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlockMortalityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlockActionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Record a flock mortality",
        "tags": [
          "flock"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/assets/{asset_id}/flock/sales": {
      "post": {
        "description": "Decrements the flock headcount by `quantity` head and books a paired INCOME event for `amount`. Rejected with 409 if it would drive the headcount below zero.",
        "operationId": "flock_create_sale",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlockSaleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlockActionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Record a flock sale",
        "tags": [
          "flock"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/material-purchases": {
      "get": {
        "operationId": "material-purchases_list_material_purchases",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "description": "Search notes and supplier",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search notes and supplier",
              "title": "Q"
            }
          },
          {
            "description": "e.g. -occurred_at,amount",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. -occurred_at,amount",
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "material_asset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MaterialPurchaseRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List material purchases in a farm",
        "tags": [
          "material-purchases"
        ]
      },
      "post": {
        "description": "Atomically records the purchase, increments the material's stock, and books an expense for the amount paid \u2014 via a paired INVENTORY and EXPENSE event. The purchased `unit` must match how the material is already tracked (any unit is allowed for a material with no inventory history). Replaying an `idempotency_key` returns the original record with status 200.",
        "operationId": "material-purchases_create_material_purchase",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialPurchaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPurchaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Record a material purchase",
        "tags": [
          "material-purchases"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/material-purchases/{purchase_id}": {
      "delete": {
        "description": "Hard-deletes the record and reverses both the stock increment and the expense. Rejected with 409 if reversing the stock would go negative.",
        "operationId": "material-purchases_delete_material_purchase",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "title": "Purchase Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete a material purchase",
        "tags": [
          "material-purchases"
        ]
      },
      "get": {
        "operationId": "material-purchases_get_material_purchase",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "title": "Purchase Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPurchaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get one material purchase",
        "tags": [
          "material-purchases"
        ]
      },
      "patch": {
        "description": "`material_asset_id` is immutable. Changing `quantity`/`unit`/`occurred_at` reconciles the inventory event; `amount`/`occurred_at` reconciles the expense. An edit that would drive stock negative is rejected with 409.",
        "operationId": "material-purchases_update_material_purchase",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "title": "Purchase Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialPurchaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPurchaseRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update a material purchase",
        "tags": [
          "material-purchases"
        ]
      }
    },
...
  }
}